Add doBlurThumbnail.js
