+++ app/scripts.babel/main.js
@@ -12,6 +12,8 @@ import stringToArray from '@functions/stringToArray';
 import formatDate from '@functions/formatDate';
 import observeAddingTalkArea from '@functions/observeAddingTalkArea';
 
+import doBlurThumbnail from '@functions/actions/doBlurThumbnail';
+
 import FileTypes from '@enums/FileTypes';
 import FormTypes from '@enums/FormTypes';
 import KeyTypes from '@enums/KeyTypes';
@@ -371,27 +373,6 @@ import settingData from '@constants/settingData';
       .filter(key => settings[key] === true)
       .map(key => SETTINGS_KEY_ACTIONS[key])
       .forEach(action => action());
-  }
-
-  /**
-	* サムネイル画像をぼかす機能を実行します。
-	*/
-  async function doBlurThumbnail(){
-    const settings = await fetchSettings();
-
-    //トークエリアの追加を監視
-    observeAddingTalkArea(talkArea => {
-      //メッセージの追加を監視
-      TalkArea.of(talkArea).observeAddingMessageArea(messageArea => {
-        const messageType = MessageArea.of(messageArea).messageType;
-        const messageHasFile = messageType == MessageTypes.FILE || messageType == MessageTypes.FILE_AND_TEXT;
-        if(messageHasFile){
-          const $thumbnailArea = $(messageArea).find('.msg-text-contained-thumb');
-          const $thumbnails = $thumbnailArea.find('img');
-          $thumbnails.each((i, thumbnail) => $(thumbnail).css('filter', `blur(${settings.thumbnail_blur_grade}px)`));
-        }
-      });
-    });
   }
 
   /**
